requirements.txt
yaml
face-recognition
streamlit_webrtc
av
opencv-python
streamlit
numpy
requests
pillow
os